+++ src/lib/config.ts
@@ -14,11 +14,7 @@ export const ServerList = [
   /**
    * API server
    */
-  "https://api.syllable.ai",
-  /**
-   * Trial API server
-   */
-  "https://trial.api.syllable.ai",
+  "https://api.syllable.cloud",
 ] as const;
 
 export type SDKOptions = {
@@ -61,8 +57,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "0.0.2",
-  sdkVersion: "0.1.0-alpha.38",
-  genVersion: "2.503.2",
+  sdkVersion: "0.1.0-alpha.39",
+  genVersion: "2.506.0",
   userAgent:
-    "speakeasy-sdk/typescript 0.1.0-alpha.38 2.503.2 0.0.2 syllable-sdk",
+    "speakeasy-sdk/typescript 0.1.0-alpha.39 2.506.0 0.0.2 syllable-sdk",
 } as const;